[types-first] Add a context flag for types-first phase
Summary:
Currently there's no way to know during typechecking whether the AST being checked is a full file or a types-first signature. This diff adds a `phase` flag to the context that records that information. When types-first is enabled, the `Merging` phase indicates that the context is being used to check a signature AST, while the `Checking` phase indicates that a whole file is being checked.

When the classic architecture is being used, the phase is always `Checking`.

Reviewed By: panagosg7

Differential Revision: D16174152

fbshipit-source-id: 1c8dcdd94b6004d82330965d720baaf01b92f315