decrease timer for slow check prints to 5 seconds
Summary: Prior to `types-first` we would print a slow merge message whenever a component took longer than 15 seconds to check. In almost all cases, this was because the component was a large cycle; it was very rare for a single file to take that long to check. In `types-first` a component now always means a single file, so using a 15 second timer for slow checks was no longer a useful metric because it never printed anything. This lowers it to 5 so we can still use this as a gauge for slow files.

Reviewed By: gabelevi

Differential Revision: D16228410

fbshipit-source-id: bc3e477a407ae3eebd84be3ee3463da8a1ad01cc