Tests for dump types and type-at-pos with abstract locations
Summary:
These already work. There are some calls to `ALoc.to_loc_exn` on the codepaths that these commands use, but it looks like `to_loc_exn` is only called on locations in the typed AST, and not on external locations that the typed AST references. Since these are all local to the file being typechecked, they always have a concrete representation and therefore `to_loc_exn` always succeeds.

It would be good to clean this up later, but it doesn't have to block launch.

Reviewed By: panagosg7

Differential Revision: D16081985

fbshipit-source-id: 0d496a1fd1b579cba06c751e87803d15f1405538