[parser] disallow await parameter in async functions
Summary:
`await` cannot be used as a parameter inside an async function.

https://tc39.es/ecma262/#prod-AsyncFunctionDeclaration

Note: the confusing part is that `FormalParameters[+Await]` means that `await` can NOT be used, rather than meaning it's all otherwise-valid formal parameters plus `await`.

also confusing, the `~await` param in `function_params` means that `await` is a keyword, meaning it is NOT a valid identifier. you might expect that it's saying that `await` is allowed as a param but nope!

Fixes https://github.com/facebook/flow/issues/7865

Changelog: fix(parser): `await` can no longer erroneously be used as a parameter in async functions

Reviewed By: gabelevi

Differential Revision: D16264960

fbshipit-source-id: 0342149879b130a726202fed97d4a89934fa0d74