[autofix][exports] command inserts generalized type for all signature verification errors in file.
Summary: This I have implemented a new autofix command that looks through the signature verification errors in a file and uses the insert-type code to insert the generalized type to fix the error. This is just a sketch of the final solution, in subsequent diffs I would like to call the type checker after insertion to see if the inserted type results in an error.

Reviewed By: mvitousek

Differential Revision: D16188351

fbshipit-source-id: 26fb06de936b14d5ab10059b1af4c01721a1ccaf