allow constant folding when the same element is spreaded twice in an array
Summary: Previously, the `ConstFoldExpansion` guard only considered two things when determining whether or not we were in an infinite recursion case: the `id` of the resolve operation, and the `reason` the of the lower bound being resolved. This was sufficient to catch cases where we would enter infinite recursion, but had the unfortunate side effect of de-opting into the generalized array case whenever the same element was spreaded into an array twice (like `[...x, ...x]`, since `x` has the same reason each time, and the `id` of the resolve operation is the same as well. To counteract this, we can also consider the position of the lower bound in the array being resolved (here proxied using the length of the list of already resolved elements). This way, these two `x`s will not be considered the same for the purposes of the `ConstFoldExpansion` guard.

Reviewed By: gabelevi

Differential Revision: D16110694

fbshipit-source-id: fa1e2056fdabedd2333eee893410495b3bfad398