[autofix][insert-type] get rid of unions of identical type annotations
Summary:
This diff moves union simplification after generalization / specialization so that it catches when two literal types are generalized to the same type.

Example:
-`Union (Num (Some "1")), (Num (Some "2")), []` generalizes to `number | number` which now simplifies to `number`
-`Union (Num (Some "1")), (Num (Some "2")), []` specializes to `Union (NumLit "1"), (NumLit "2"), []` which doesn't (and shouldn't simplify)

Generalization and specialization are heuristics that we apply to types that would otherwise serialize to temporary types. I have hopes of one day producing a polarity tracking version so that we can suggest the most permissive sound type possible (I think it would be generalizing in contravariant positions and specializing in covariant positions) but this would require much more effort, because you actually need a map that that tracks the variance in the AST and Ty which sounds like a lot of work.

By simplification I mean the process of removing redundant type information in unions and intersection in a hopefully sound manner. This diff improves simplification in the following two ways.

- Using the ty_simplifier with a special comparison operation that allows different types of Any to be considered equal during simplification.
   - `Union (Any Implicit) (Any Explicit)` simplifies to `Any (Implicit | Explicit)` because both would have been serialized to the same ast node.
- It also adds a heuristic to remove literal types from unions when the base type is present.
   - The types `number | 1 ` and `"foo" | string` simplify to `number` and `string` respectively because the literal types are subtypes of the base types,
   - `true | false` simplifies to `boolean` because this should be semantically equivalent, and
   - `string | 2 | 3` doesn't simplify to `string | number` because doing so wouldn't preserve the meaning of the original type.

Reviewed By: panagosg7

Differential Revision: D16136811

fbshipit-source-id: 595cff99c9bd089cfdb5736d24a3e0c36775e95c