Adding use_ops to Resolved and FullyResolved
Summary:
When a constraint is resolved, its `use_op` is lost and a bad error is likely to happen.
For example, in tests/badly_positioned_spread/bad.js, the definition at 10 causes the error,
but the imported definition at foo.js is blamed.

In this diff, `Resolved` and `FullyResolved` preserve the use_ops.
These use_ops are used when a corresponding `OpenT` is used as an upper bound in `__flow`.

Reviewed By: avikchaudhuri

Differential Revision: D16352931

fbshipit-source-id: 289b2d3d9c06f7295cf8e60b247b291fec3f3f54