[hack] Support loading a specified naming table saved-state from disk
Summary:
Support loading a specified naming table saved-state from disk. This is useful for testing, but could also be useful for open-source users if they want to opt into the serverless IDE. (Note that it's currently not open-source, but we could change that after this commit.)

This commit adds support for specifying the path to the naming table saved-state on disk. It also causes the serverless IDE to wait for the naming table to be loaded rather than immediately failing, which is useful for testing. (Otherwise, we would have to `sleep` and hope that the serverless IDE initialized).

A naming table saved-state can be generated by running `hh_server --check . --save-naming path/to/naming.sqlite3`, which can then be passed into the initialization options.

Reviewed By: jewelpit

Differential Revision: D16217540

fbshipit-source-id: a87f81e94595f72a949bd99a1b30405823d88ce8