[def-inst-prop-assign] add type annotations to tests
Summary: We need to add type annotations to the properties in our tests in preparation for the voidable analysis. Once the voidable analysis is turned on, many of the existing tests would no longer have errors since in my tests, most properties lacking a type annotation happen to be voidable. I will add more tests throughout the stack to explictly test that the voidable analysis is working without having to ruin our existing tests.

Reviewed By: mroch

Differential Revision: D16366165

fbshipit-source-id: 4d7299199d1b8d4c2b65ca3bee3cd99eb6b20295