Fixing a bad error related to object_kit
Summary:
For the newly added test, an error is caused by the use at line 24.
The definition at line 6, however, used to be blamed.

The root of this bad error is the `unknown_use` introduced by object_kit.
There was a comment indicating that the `unknown_use` was used intentionally,
though replacing the `unknown_use` does not break any existing test.

Reviewed By: panagosg7

Differential Revision: D16441534

fbshipit-source-id: 7638cccceab24c836910ff1f97240d01e437eba4