[hack] Add optional `powered_by` field to JSONRPC responses
Summary:
This is a non-standard extension to the JSONRPC protocol. It adds the `powered_by` field as an optional field to responses and notifications, which may be used for logging or as a diagnostic aid.

This is useful for the serverless IDE to indicate which back-end a given request was served by.

We call it `powered_by` in symmetry with the somewhat well-known `X-Powered-By` HTTP header.

I did not find any existing extensions to JSONRPC to base this off of.

Reviewed By: jewelpit

Differential Revision: D16370668

fbshipit-source-id: a1eb1ff58ac0f8bb497a9acb49da04e6266d8572