[normalizer] Remove use of ALoc.of_loc_exn
Summary:
`find_type_at_pos_annotation` is useful when the input location might be
within a span that appears in the typed AST. In `extract_scheme` the locations
are extracted from the AST itself (when the file_sig is populated) this means that
we can use `find_exact_match_annotation` instead, that matches whole spans
rather than check for inclusion. This also does not need to use the unsafe
`ALoc.to_loc_exn` conversion since matching is done at the `ALoc.t` level.

Reviewed By: dsainati1

Differential Revision: D16471918

fbshipit-source-id: 4c8e81f926e1d9dae2ce257ddf03f4b08a01581e