[hack] Enable ocamlformat for parts of hphp/hack
Summary:
Enable `ocamlformat` for `hphp/hack`.

Since the codebase is large and not consistently formatted today, only format the AST and AAST directories. For the others, leave a `.ocamlformat` that disables formatting in that directory. This also ensures ensures that any new directories will be formatted in this style.

Configure using a `.ocamlformat` that's close the current style to minimise code churn.

Reviewed By: jewelpit

Differential Revision: D16421601

fbshipit-source-id: 7c71aa5fd295f6163a63fd10112da2042c1234ad