[PR] Bump required ocaml version to 4.07.1
Summary:
We're bumping the required ocaml version to 4.07.1, in order to stay up to date. There's no internal testing against other versions, so although at this moment we're currently technically compatible with 4.05.0, we are bumping the minimum "required" version to 4.07.1. It may work on older versions but it is not supported.
Pull Request resolved: https://github.com/facebook/flow/pull/7958

Reviewed By: gkz

Differential Revision: D16429747

Pulled By: mroch

fbshipit-source-id: 8408f7333c5ae28c283fca34feab3f322eadcf11