use locations as ids for object types created from source
Summary:
Comparing objects in Flow is very slow, so we like to avoid doing it as much as possible. To this end, when we see two objects with the same property map `id`, we take the "fast" path and skip comparing their fields, since we know for certain that they must be equal.

In types-first mode, however, we can cause these `id`s to be superficially different by exporting and re-importing a type in a cycle. Consider the following code:

A.js:

```
const b = require('B.js')
export type Foo = { ... }
(b : Foo);

```

B.js

```
import type {Foo} from 'A.js';
declare var b : Foo;
module.exports  = b;
```

During merge, we produce an object type for `Foo` with `id1`, and use that to determine the type for `b` when it is exported from `B.js`. However, during the check phase, when checking `Foo`, we produce a new object type with `id2`. Then, when flowing `typeof b ~> Foo`, we have two instances of the same object, with differing ids. If `Foo` is some large, complicated object type containing spreads or unions or other such slow features, this means Flow has to do a bunch of redundant work just to decide that `Foo <: Foo`.

We can avoid this by using locations as the id for object types, since they will be the same in the check and merge phases. However, we don't always create objects based on locations, so we represent an object id with either an int (as before) or a location. At the moment we only use locations for object types from source, but we can almost certainly do this in more places than just that.

Note that we only use locations as ids for object types that appear at the "top level" of a file, that is, in Global or Module scope. This prevents us from having id collisions when traversing code multiple times, like in `generate_tests` for example.

Reviewed By: samwgoldman

Differential Revision: D16346737

fbshipit-source-id: 87de021dd28d330fc7dab42e5b3f044d92b94990