abstractify locations before using them as object ids
Summary:
A previous change updated object types to use locations as their property ids, rather than integers, when they come from source code. The effect of this is to prevent us from doing unnecessary work to compare equivalent objects.

Abstract locations complicates this change by introducing two types of locations that need to be handled, a concrete location (produced during the check phase) and an abstract location (produced during the merge phase). When a cycle causes a file to import an object type exported in its own file, we can encounter a scenario where the object's id when seen in merge is abstract, and when seen in check is concrete. This results in the cache failing to recognize that these types are the same.

To solve this, we only use abstract locations as object ids, even in the check phase. When producing an object type from source in the check phase, we look up the abstract equivalent to the concrete location of the object type, and use this as the id of the object. Because we look up this info using the inverse of the location tables used in the merge phase, we guarantee that this is the same id that the object had during merge. This way, the cache is still hit and we can skip any check comparing them.

Reviewed By: nmote

Differential Revision: D16469791

fbshipit-source-id: de3768fe53f76c186b4d3063f2ee133c07f51108