[ty_simplifier] allow sorting of unions and intersections
Summary:
This diff allows the `ty_simplifier` to be used for sorting types. To enable this I altered `Nel.dedup` to guarantee the returned `Nel.t` is sorted. Doing so enables us to remove the iteration on pointer equality at the topmost `Ty.t`.

This will be used in a follow-up diff to make types suggested by insert-type more readable, such as putting `any`s and literal types at the front of unions.

Reviewed By: panagosg7

Differential Revision: D16345144

fbshipit-source-id: 70992f5fad076ed816e1346e9249fe98f400f967