[autofix][insert-type] enable specialization to return literal types and soundly simplify union types
Summary:
Generalization and specialization are heuristics that we apply to types that would otherwise serialize to temporary types. I have hopes of one day producing a polarity tracking version so that we can suggest the most permissive sound type possible (I think it would be generalizing in contravariant positions and specializing in covariant positions) but this would require much more effort, because you actually need a map that that tracks the variance in the AST and Ty which sounds like a lot of work.

By simplification I mean the process of removing redundant type information in unions and intersection in a hopefully sound manner (much like the ty_simplifier). This diff adds an extra pass over the Ty.t generated in insert-type to remove type literals from unions when their base type is present in the union, or when the union enumerates all literal elements (only possible for booleans). This could eventually be added to the Ty_simplifier, but I don't think that it necessarily has to be added there.

- It removes literal types from unions when the base type is present. The types `number | 1 ` and `"foo" | string` simplify to `number` and `string` respectively because the literal types are subtypes of the base types.
 - `string | 2 | 3` doesn't simplify to `string | number` because doing so wouldn't preserve the meaning of the original type.
 - `true | false` simplifies to `boolean` because this should be semantically equivalent

Reviewed By: panagosg7

Differential Revision: D16345140

fbshipit-source-id: 9c9ec5e509cc2071a053d695cdcad82fcdf4c3ef