Clean up watchman timeouts
Summary:
I wanted to specify no initialization timeout for {D16358621} and one thing led to another and I ended up cleaning up how timeouts are specified in `watchman.ml`. The general idea is that a timeout is either

* `No_timeout` - We shouldn't timeout. Previously we used `0.`, since `Unix.select` will wait forever when passed that, though not all our timeout code supported that.
* `Explicit_timeout time_in_secs` - Timeout after `time_in_secs` seconds
* `Default_timeout` - The default is hardcoded to 120.

I think this is more clear than trying to do it with `float option`.

Reviewed By: mroch

Differential Revision: D16518447

fbshipit-source-id: 81cd6fd601ef0213904a78a7086bb8324112a013