Add `--types` flag to `flow graph dep-graph`
Summary:
Type dependencies are different than value ones. For example the import of `foo` in
```
// file b.js
import {foo} from 'a';
foo();
module.exports = class A {};
```
does not affect the type signature of file a.js since it does not affect its exports' type.

The edge
```
a.js --> b.js
```
is only a value dependency but not a type one.

`flow graph dep-graph` used to only consider value dependencies. This diff adds an
option `--types` to only account for type dependencies.

This flags makes a difference in types-first mode.

Reviewed By: avikchaudhuri

Differential Revision: D16567073

fbshipit-source-id: 8d2b494346056864e6ebaa768a628e856b30b5ca