[spreads] Allow slices in the todo_rev list
Summary:
When we look at an object type, we look at all of the properties, the spreads in between, and the indexers in between and we interpret that as a list of spreads. For example, this:
```
type A = {a: number, ...B, c: number, [string]: number, ...D};
```
becomes :
```
type A = {...{| a: number |}, ...B, ...{| c: number, [string]: number |}, ...D};
```

Internally, this becomes a TypeDestructorT of SpreadType, with each intermediate object in a list. When we evaluate the spread, we resolve each object and turn it into a `slice`, which is the representation we use to compute the spread.

In type_annotation.ml, we create this list of objects to resolve by looking at the properties/spreads and creating a list of Spreads and Slices. We then take that list of Spreads and Slices and turn them all into object types and put them in the list for the SpreadType destructor to handle.

Instead of converting the slices into objects and then back into slices, we can just keep the slices around directly and pass them into the spread computation. This saves us some work, since we don't need to convert the slice to an object and then back into a slice.

There were a few extra things we needed to handle with this change. Since `slice` used to only be accessible via a `use_t`, certain classes/functions that walk the structure of our types never had to handle them. This diff also has the changes necessary to make those work.

Reviewed By: samwgoldman

Differential Revision: D16039088

fbshipit-source-id: cc2d7128e22a1c0cec6c546aade651444b7cd526