[spreads] Error when indexer may be overwritten by inexact, add type_spread tests, fix existing tests
Summary:
If the object on the right is inexact and the object on the left has an indexer, it's possible for the object on the right to add properties that are incompatible with the indexer.

This also finishes up the implementation of type spread, so I've fixed existing tests and added a few more.  The `type_spread` tests exercises every case in the design. I separated it out from `new_spread` to make reviewing easier.

Reviewed By: samwgoldman

Differential Revision: D14016367

fbshipit-source-id: 40a9e616c48881781f3307e086acc3aed12f4800