[spreads] More precisely handle the case when a property only exists on the right
Summary:
Instead of always erroring when the object on the left is inexact and doesn't mention some property `p` on the right, we can split into two cases.

1. If p is optional and the property on the left is inexact or has an indexer (`{} + {p?: number}`) then we error, since the inexact object may contain p with another type. We could be more precise here when the object on the left has an indexer, but it's not clear that people _want_ more inferred unions.

2. Otherwise (p is not optional or the object on the left is exact with no indexer) we can just take `p` with its type.

All that remains is to error when an object on the right is inexact and the object on the left has an indexer (because the inexact object may have props that conflict with the indexer). I'll implement that check in the next diff, and then I'll fix up all the tests.

Reviewed By: samwgoldman

Differential Revision: D14015200

fbshipit-source-id: 0c94019a05a337526f30b8356a84305857adb797