[spreads] Disallow interface spreads
Summary:
Interfaces don't track the own-ness of props, so we should error when we try to spread them.

Note, I position the errors around the interface location (instead of the spread location) because polymorphic
type signatures that spread and cause errors should not be suppressed at the call, not the definition, of the function.

The test I added shows an example.

Reviewed By: samwgoldman

Differential Revision: D13969033

fbshipit-source-id: e07d3fa604113764962fb80c3a1318564ed2ee00