[spreads] Implement merge_result for calculating spreads that may error
Summary:
`merge_result` can be used to lift a binary function over spreads that may return a Flow error to a function over resolved lists. This will be useful later in the stack-- spread2 will no longer _always_ compute a spread as a result. If one of the slices are unsealed, for example, we should error instead. In order to stop the spread calculation ASAP and propagate the error, we can use a `(slice, FlowError.error_message) result` to model the return value of `spread2`.

This diff also changes the signature of `spread2` to return a `(slice, FlowError.error_message) result`, but does not alter the behavior of `spread2`. I do that to test the correctness of the `merge_result` function, and I'll need to do it later in the stack anyway.

I included a local bind function that matches the canonical `>>=` signature. This is the only time I've worked in OCaml where I wished it was Haskell instead.

Reviewed By: samwgoldman

Differential Revision: D13996424

fbshipit-source-id: 5fe645f71ee5e14128b02be00b19da8a1fa2f1af