[spreads] No indexer on right, take indexer from left
Summary: Error on indexers appearing on the right in spreads and carry over the indexer from the left. We do this because the indexer on the right may overwrite keys that already exist in the object, but we can't know which ones because we do not know what specific keys the indexer may be hiding.

Reviewed By: samwgoldman

Differential Revision: D13999335

fbshipit-source-id: 0bae6aef3402826333f8b029f69f2d2022fc7a64