[spreads] Don't resolve slices that are already resolved
Summary:
After syntactically walking a spread, we get a list of spreads and slices. Every two slices must have a spread between them, otherwise they would just be one larger slice.

This diff serves two purposes:
1. If the last element in our list of spreads and slices is a slice, we used to turn that slice into an object, resolve it, and then turn it back into a slice. Now, we just feed that slice directly into the spread mechanism instead.
2. If the last element in our list was a slice, converting it into an object makes us forget that it was originally a slice when we find it in `spread2`. Since slices have special behavior in type spread, we need to make sure we keep around this information.

I propagated this information by changing the `acc` list in the spread state to hold either an `InlineSlice` with the information from the syntactic spread, or ` ResolvedSlice` which holds the information from a resolved object that was turned into a slice.

Most of this diff is the normal set of files you need to update after changing type.ml.

Notably, this fixed a bug in the type normalizer where a nested spread after a slice was reversed. See the updated test for more information.

Reviewed By: samwgoldman

Differential Revision: D16353144

fbshipit-source-id: 2c8af596cf500428b39a72f5bde0ec906a466746