[spreads] Move object_kit to its own file
Summary:
I'm planning to reuse some of the internal functions in object_kit in order to implement spread on values. Moving it to its own module makes
it easier to expose only certain bits of code outside of the module. Also, this makes a nice dent in reducing the LOC in flow_js.ml

Reviewed By: samwgoldman

Differential Revision: D14168027

fbshipit-source-id: ef9a8744af396391b02a513425086952fa494d40