[spreads] Inline properties overwrite properties on the left, allow inline indexer on the right
Summary:
Optional properties inline should overwrite properties on the left.
```
{...{p: string}, p?: number} = {p?: number}
// Previously, this would be p: string | number
```

This may seem counter-intuitive, since
```
{...{ p: string }, ...{ p?: number}} = {p?: number | string}
```

We are making this design decision intentionally in order to reduce the amount of union types we infer. It's unlikely that a property specified inline actually was meant to be inferred as a union instead of the type specified inline.

This is also safe-- you can't specify an optional property value inline, so there's no runtime equivalent that would break here. That is to say, you can't do this at runtime
`{...{p: 'string'}, p?: 3}` because that's not valid JS code.

Similarly, we allow indexers to be specified on the right inline:
```
{...{ p: string }, [string]: number}  = {p: number, [string]: number}
// Previously, we would error here
```

It's reasonable to interpret this type in this way. Since indexers also have no runtime equivalent, this is safe to do for types.

The left hand side also does not need to be inexact for this leniency for the same reason!

Note that we still don't need to do anything special to make sure we're in a type spread once we move value spread to this mechanism. Since the only properties you can specify inline are explicit props or computed props, we will never hit these special allowances.

Reviewed By: samwgoldman

Differential Revision: D16365012

fbshipit-source-id: 584185051e3d1867a323977062ecf825ad138574