[spreads] Treat mixed as {} in spreads
Summary:
We treat `mixed` as `{[string]: mixed}` in `object_kit`, which gives `...mixed` an indexer. This leads to confusing errors, since the indexer does not explicitly exist.

You can never spread mixed as the second object in a spread anyway, since it is inexact. We can create a better error message by just treating it as an empty inexact object.

If `mixed` is spread first, this gets rid of the implicit indexer, which is safer than implicitly including it.

Reviewed By: samwgoldman

Differential Revision: D14026870

fbshipit-source-id: 625599189c973717bbc3432b933fafb0ea7950d1