[spreads] Make spreads go left to right
Summary:
Right now, `spread2` is an associative function. In a future diff, I will make non-associative by allowing inline properties/indexers on the right to overwrite things on the left. That mean we _need_ to evaluate spreads from left to right.

Coincidentally, that's consistent with how they are evaluated at runtime anyway, so this just brings us to a closer representation of what actually happens at runtime.

Before, we used to gather a list of slices to resolve by looking at the object type. That resulted in a list of slices that syntactically occurred right to left. Then we'd reverse that list and pass it as "todo_rev". After we resolved all of them, we'd reverse the list again (now back to reverse order) and then run spread on it. First we'd consider the last two elements, then that and the third to last element, etc., moving leftwards syntactically and considering the new left object and the accumulation from the right.

In this diff, we don't reverse the list to pass into todo_rev (now named todo), and then we reverse it so that it's ordered like it was in the syntax again. Then we run spread on it, but instead of moving right-to-left we move left-to-right (see change in merge_result).

This happened to also find and fix an error in the way we normalize types. See the fixed type-at-pos test case.

Reviewed By: samwgoldman

Differential Revision: D16227294

fbshipit-source-id: b89e21cc4f089f582faf709566c73bcabbefd5f1