[def-inst-prop-assign] error on bad reads in default initializers
Summary:
Previously, it was possible to read from an uninitialized property (or cause `this` to escape before all properties were initialized) in a property's default initializer. This Diff fixes that soundness bug by extending the analysis to include default property initializers and emitting the appropriate errors.

The runtime semantics for default initializers are that they are executed before the constructor in base classes and immediately after `super` is called in derived classes. For the purposes of our analysis, it suffices to treat them as if they are executed before the constructor in all cases since the analysis already establishes that all ancestor properties are initialized inductively and you cannot even refer to the properties before `super` is called.

Most of the changes are just moving things around so that we can expose the code for marking a property as initialized. An alternative implementation was to convert each property declaration with a default initializer into an assignment statement by building fake AST nodes and prepending those assignment statements to the body of the constructor. I believe the implementation I ultimately chose is superior to the alternate one because building fake AST nodes is prone to violating invariants and creates unnecessary intermediate values.

Reviewed By: jbrown215

Differential Revision: D16270444

fbshipit-source-id: dbea42c7081cf80812f675546e887aaa26b669b6