[def-inst-prop-assign] look at voidable checks post-merge
Summary:
We should only emit errors for a given property if VoidT flows to the type of that property. Ideally, we would create a VoidT ~> property type flow in `statement.ml` for each potential error. The problem wuth adding a flow from void is that it affects other constraints on the property's tvar which causes weird type inference behavior, such as emitting spurious errors even when the lint is off. After talking with Sam, we concluded that type inference behavior should never depend on lint settings, thereby ruling out my initial implementation. The solution we settled on is collecting all of potential errors that we would have created a flow for in the context and dealing with them post-merge. At this point, the Tvar for a property will _probably_ have a concrete type. If the tvar has a concrete type, we can simulate flowing void to a property by pattern matching on the property's type. If the tvar is still unresolved post-merge, we conservatively consider that property to be non-voidable and emit errors for it. In the future, we could potentially do more checks when the tvar is unresolved to try to reduce the number of spurious errors.

(We start adding `voidable_check`s to the context in the next Diff. In this Diff, I just want to focus on the decision to perform the voidable analysis post-merge.)

Reviewed By: jbrown215

Differential Revision: D16498025

fbshipit-source-id: 29f2459b0271ee7aada7c9dc0384baeaaf186b9a