[def-inst-prop-assign] refactoring the analysis
Summary: There refactor primarily consists of three changes: (1) putting the suberror type in `lints.ml`, (2) passing the entire class body to `Property_assignment.eval_property_assignment`, (3) and changing the return type of `Property_assignment` so that all of the suberrors are in one unified structure. I chose to do a refactor now because it will simplify the changes in the rest of the stack.

Reviewed By: mroch

Differential Revision: D16497931

fbshipit-source-id: 3ec8ff38cd1a0c47609d074531a784e49f2dc2a0