[def-inst-prop-assign] don't error inside property function defs in ctor
Summary:
This Diff causes us to not error in the body of a property function and instead error if that function is called before all properties are initialized. With this change, we now treat instance property functions and methods the same way for the purposes of our analysis.

We currently analyze the body of a function defined in the constructor at the point where it is defined. Ideally, we would wait to emit any errors until the function is called, as that is when the questionable code is actually executed. Fixing this imprecison for all functions defined in the constructor (property functions, let-bound functions, IIFEs, etc.) is a relatively tricky problem that would require additional bookkeeping for all anonymous functions created in the constructor, but we can easily fix this problem for property functions specifically because of all the work we already do to prevent `this` from escaping the constructor early. Namely, the same logic that we currently use to prevent methods from being called before all properties are initialized already works for property functions too.

This fix is not just theoretical-- a common JavaScript idiom is to assign arrow functions to instance properties. (TIL this is often better than creating an instance method because invoking a method on an instance does not automatically bind `this` to that instance.) Therefore, this fix should decrease the number of spurious real-world errors emitted by my analysis.

Conveniently, this also fixes another KP with defining recursive and mutually recursive property functions. The usual order for evaluating assignment expressions for properties, i.e. `this.<property-name> = <expr>`, is to read the RHS and then write to the LHS. We do this to prevent nonsensical assignments like assigning a property to itself or `this.p = this.p + 1` (where `this.p` is previously uninitialized). Clearly, this evaluation order does not work for properties that are recursive functions since we will get a `ReadFromUninitializedProperty` error when evaluating the RHS. This Diff solves that evaluation order problem by not evaluating the RHS at all for properties that are assigned to functions (which is sound for the reasons explained in the above paragraph).

Reviewed By: jbrown215

Differential Revision: D16292465

fbshipit-source-id: 163e29b9d5216d89439a43efe9dc4cf8de182e93