[def-inst-prop-assign] reads from voidable properties
Summary: This Diff allows you to read from an uninitialized, voidable property. The idea is that if a property is voidable, the programmer already has to handle the possibility that it is uninitialized at runtime. Before this Diff we would never allow a read from an uninitialized property regardless of its voidability.

Reviewed By: jbrown215

Differential Revision: D16366237

fbshipit-source-id: dc17ffe6f7cd705a9aa28363a8f765da7a1938d9