[hack] Fix SharedMem.heap_size
Summary:
The OCaml implementation of SharedMem.heap_size is subtly incorrect which can
lead to overcounting.

In order to avoid counting the same value twice, the traversal maintains a seen
set of block value addresses. However, this function allocates, and the
allocation can trigger GC which can move values around. This would confuse the
algorithm and cause it to repeatedly traverse these values.

The Obj module provides a function for this which is implemented in C which does
not suffer from this problem. As in extern.c, the seen values are tracked by
setting the GC bits in the header.

The Obj implementation also avoids traversing statically-allocated data, which
poses a small complication for testing. I've inserted a simple hack to ensure
test data is in the heap.

Reviewed By: gabelevi

Differential Revision: D16643321

fbshipit-source-id: 664ed9ddde0a3dbadc5e143adae31ab54508ddbe