[enums] Enum member names must not start with [a-z]
Summary:
Rationale: Enum methods start with lowercase ‘a’ through ‘z’. This way an enum member name can never shadow a method, and there is never any confusion on the part of the user.

Additionally, all reserved words start with a lowercase [a-z], so in effect we have banned reserved words from being used as enum member names. I have removed a test associated with this as it is no longer relevant.

Reviewed By: gabelevi

Differential Revision: D16677274

fbshipit-source-id: 31ad225474ea0430d12482de522db9d97d0632fd