DirectDependentFilesCache
Summary:
The `DirectDependentFilesCheck` step is an O(repo) sized check that runs on every recheck (except unchecked dependencies). For small rechecks, this is the slowest step.

If we memoize the result of `DirectDependentFiles` and only throw it away when the dependency graph changes (i.e. new resolved requires are written to the heap) then we can make tiny rechecks ~60% faster. If a later diff does the same with `AllDependentFiles` then we probably can make tiny rechecks ~85% faster.

Reviewed By: avikchaudhuri

Differential Revision: D16447249

fbshipit-source-id: 49a2126a3cabf6d7bdae9d499252b61d59064e16