Track if resolved requires change
Summary:
At the moment this only triggers some logging. It will either log `Resolved requires changed` or `Resolved requires are unchanged` after we write resolved requires to the heap.

This happens during the `ResolveRequires` step, when we write the resolved requires for the changed files.

This also happens during the `ReresolveDirectDependents` step, where we write the resolved requires for the direct dependents of changed files.

Reviewed By: avikchaudhuri

Differential Revision: D16034453

fbshipit-source-id: cae5f47e9a293799d8ffb41f7dfe339317677215