Hash resolved requires
Summary:
We represent the dependency graph as the resolved requires heap. I want to be able to tell when the resolved requires heap changes. In order to do this, we need an easy way to tell if the new resolved requires we're writing is different than the previous resolved requires.

We can do this easily by keeping a hash for each resolved requires.

Reviewed By: avikchaudhuri

Differential Revision: D16010583

fbshipit-source-id: da180cc8a22b9b4a8647298f9b7a482751d9e6dc