fix SSA for missing catch
Summary:
A missing `catch` in a `try-finally` statement would be mistakenly handled similarly as an empty `catch` in a `try-catch-finally` statement.

This had the effect of clearing any abrupt completion in `try` before entering `finally` in the `try-finally` statement. Assuming the `finally` itself exited normally, we would essentially "forget" the abrupt completion in `try`.

Instead, we now do not clear an abrupt completion in `try` before entering `finally` (when `catch` is missing), and that abrupt completion is "rethrown" after the `finally` exits.

Reviewed By: gabelevi

Differential Revision: D16751747

fbshipit-source-id: 28652eefe769caa0f7540f5fbd6b6d356f1a4cea