[Ty] Clean-up the Ty_utils simplify API
Summary:
One of the main utilities on `Ty.t` is the one that simplifies types. There are three
kinds of transformations that are applied
1. remove `empty` (resp. `mixed`) types from unions (resp. intersections).
2. convert intersections (resp. unions) that include `empty` (resp. `mixed`) to `empty` (resp. `mixed`).
3. remove duplicates from unions/intersections.

For (3) the Ty utility uses a specialized comparator class. This allows us to determine some
types that we want to consider the same under comparison. For example, under certain modes of
operation we want to `Any Implicit` and `Any Explicit` to be considered the same. (Similarly with
`Bot Annotation` and `Bot (Unsafe _)`.)

This diff introduces a flag `merge_kinds` for `simplify_type` that unifies this behavior for both
`empty` and `any`. (Before the flag used to only cover `empty`.)

The rest of the diff propagates this change and restructures the respective tests. Note that
some of the internals of this process spilled out of Ty_utils. This has been restored so that
the only way to interact with simplification is through this flag.

Reviewed By: samwgoldman

Differential Revision: D16723964

fbshipit-source-id: bdb35aad4072219f5bcc51c7803409869915b5b0