fix off-by-1 error in traces
Summary:
Spec:
* `--traces 0` (default) prints no traces
* `--traces 1` prints traces that have no (or cut off) parents
* `--traces 2` prints traces that have no (or cut off) grandparents
* etc.

Previously, even if we had `--traces N` where `N` was larger than the maximum depth of a trace, we'd still see some grand*parents cut off in the trace. This is because we print traces up to a `level` that is calculated to be the minimum of `N` and the `max_depth` of a trace, but the `max_depth` was calculated to be `0` for a trace with no parents, `1` for a trace with no grandparents, etc. In other words, if `N` was less than the maximum depth of a trace, we'd have expected behavior, but otherwise we'd cut off one more level than expected.

Reviewed By: samwgoldman

Differential Revision: D16840772

fbshipit-source-id: 6578e4f54f8b0159d721b1ac1983a3515fa5f93c